Move Tag.js to commons/ui and removed useEffect from TagBox.js
