+++ mwdb/web/src/components/RecentBlobs.js
@@ -5,7 +5,7 @@ import queryString from "query-string";
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 
 import { RecentView, RecentRow, RecentInnerRow } from "./RecentView";
-import { TagList } from './Tag';
+import { TagList } from "@mwdb-web/commons/ui";
 import { DateString, ObjectLink, Hash } from "@mwdb-web/commons/ui";
 
 export function RecentBlobRow(props) {
