+++ mwdb/web/src/components/RecentConfigs.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { RecentView, RecentRow, RecentInnerRow } from "./RecentView";
-import { TagList } from './Tag';
+import { TagList } from "@mwdb-web/commons/ui";
 
 import { DateString, ObjectLink } from "@mwdb-web/commons/ui";
